refactor: Separate Android application into a dedicated :androidApp module
- Migrated Android application logic, manifest, and resources from `:composeApp` to a new `:androidApp` module.
- Reconfigured `:composeApp` as an Android Multiplatform Library using `com.android.kotlin.multiplatform.library`.
- Updated `compileSdk` to 37 in `libs.versions.toml`.
- Refactored `MainActivity` to use `@PreviewLightDark` for simplified theme previewing.
- Cleaned up `gradle.properties` by removing redundant Android build flags.
- Updated project structure to include `:androidApp` in `settings.gradle.kts`.
