+++ androidApp/src/main/kotlin/io/dimasla4ee/shoppinglist/MainActivity.kt
@@ -1,12 +1,13 @@
-package io.dimasla4ee.shoppinglist.app
+package io.dimasla4ee.shoppinglist
 
-import android.content.res.Configuration
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.tooling.preview.PreviewLightDark
+import io.dimasla4ee.shoppinglist.app.App
 import io.dimasla4ee.shoppinglist.app.ui.theme.ShoppingListTheme
 
 class MainActivity : ComponentActivity() {
@@ -20,8 +21,8 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Preview("Light Theme", showSystemUi = true)
-@Preview("Dark Theme", showSystemUi = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Preview
+@PreviewLightDark
 @Composable
 private fun AppAndroidPreview() {
     ShoppingListTheme {
