+++ androidApp/src/main/res/drawable/ic_launcher_monochrome.xml
