+++ settings.gradle.kts
@@ -33,3 +33,4 @@ plugins {
 }
 
 include(":composeApp")
+include(":androidApp")
